Check whether cm_requestClient still exists!

+++ resources/db/update/37.php
@@ -10,6 +10,8 @@ if (!CM_Db_Db::existsTable('cm_requestClientCounter')) {
     CM_Db_Db::insert('cm_requestClientCounter', array('counter' => 0));
 }
 
-$highestEntry = (int) CM_Db_Db::select('cm_requestClient', 'id', null, array('id' => 'DESC'))->fetchColumn();
-CM_Db_Db::update('cm_requestClientCounter', array('counter' => $highestEntry));
-CM_Db_Db::exec('DROP TABLE IF EXISTS `cm_requestClient`;');
+if (CM_Db_Db::existsTable('cm_requestClient')) {
+    $highestEntry = (int) CM_Db_Db::select('cm_requestClient', 'id', null, array('id' => 'DESC'))->fetchColumn();
+    CM_Db_Db::update('cm_requestClientCounter', array('counter' => $highestEntry));
+    CM_Db_Db::exec('DROP TABLE IF EXISTS `cm_requestClient`;');
+}
